migrate some method out of managed ledger
